Rename files to reflect the mvc pattern I inadvertently used
