+++ gui/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-    slint_build::compile("ui/app.slint").expect("Slint build failed");
+    slint_build::compile("view/app.slint").expect("Slint build failed");
 }
