+++ gui/src/model.rs
