+++ gui/view/app.slint
